% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codigos_departamentos_pe}
\alias{codigos_departamentos_pe}
\title{Códigos y Acrónimos de Departamentos del Perú (Lamas y Encarnación, 1976)}
\format{
Un data frame con 24 observaciones y 3 variables:
\describe{
\item{acronimo}{Código de 2 letras en mayúsculas (ej. \code{"AM"}, \code{"CU"}, \code{"LO"}).}
\item{departamento}{Nombre completo del departamento o región del Perú.}
\item{ubigeo_inei}{Código de ubicación geográfica departamental de 2 dígitos según el INEI.}
}
}
\description{
Tabla de referencia con los acrónimos estándar de dos letras utilizados
históricamente en biogeografía peruana y adoptados por el MINAM en los listados
oficiales de flora CITES para registrar la distribución geográfica departamental.
}
\details{
En las tablas de distribución botánica del MINAM, se utiliza además el símbolo especial
\code{"?"} para señalar que una especie cuenta con registro o colecta confirmada en el Perú,
pero el espécimen de herbario carece de localidad o departamento explícito en su etiqueta.
}
\examples{
data(codigos_departamentos_pe)
head(codigos_departamentos_pe)

}
\references{
Lamas, G. & Encarnación, J. (1976). Acrónimos departamentales para registros biogeográficos peruanos.
}
\keyword{Peru}
\keyword{datasets}
\keyword{departamentos}
\keyword{geografia}
