% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citesperu-package.R
\docType{package}
\name{citesperu-package}
\alias{citesperu}
\alias{citesperu-package}
\title{Peruvian Species Checklists from CITES Publications}
\description{
An R package under development for working with fauna and flora checklists
published by Peru's Ministry of the Environment (MINAM). This initial version
documents the project; datasets and query functions are not yet included.
}
\details{
Planned features include reproducible data preparation, checklist queries,
scientific name matching and summaries by appendix and taxonomic group.
Each future dataset will identify its source and edition independently.
}
\seealso{
\url{https://www.gob.pe/institucion/minam/colecciones/609-listados-de-especies-de-fauna-y-flora-cites-peru}
}
\author{
\strong{Maintainer}: Paul E. Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Paul E. Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID}) [copyright holder]
}

}
\keyword{internal}
